# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AdBoard API",
    "version": "v1"
  },
  "paths": {
...
    "/Users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Вернуть всех пользователей",
        "responses": {
          "500": {
            "description": "Произошла внутрення ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "Запрос выполнен успешно",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShortUserDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShortUserDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Получить пользователя по id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор пользователя",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Произошла внутрення ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "Запрос выполнен успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "404": {
            "description": "Пользователь с таким id не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Обновляет пользователя по модели",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор обновляемого пользователя",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "Login",
                  "Name",
                  "Password"
                ],
                "type": "object",
                "properties": {
                  "Name": {
                    "maxLength": 64,
                    "minLength": 4,
                    "type": "string",
                    "description": "Имя пользователя"
                  },
                  "Login": {
                    "maxLength": 64,
                    "minLength": 4,
                    "type": "string",
                    "description": "Логин"
                  },
                  "Password": {
                    "maxLength": 50,
                    "minLength": 8,
                    "type": "string",
                    "description": "Пароль"
                  }
                }
              },
              "encoding": {
                "Name": {
                  "style": "form"
                },
                "Login": {
                  "style": "form"
                },
                "Password": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Произошла внутрення ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "Запрос успешно выполнен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShortUserDto"
                }
              }
            }
          },
          "400": {
            "description": "Модель невалидна",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Пользователь не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "Конфликт бизнес логики",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Удаляет пользователя по ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор пользователя",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Произошла внутрення ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "204": {
            "description": "Удаление произведено успешно"
          },
          "403": {
            "description": "Нет доступа",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Пользователь с таким ID не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      }
    },
    "/Users/current": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Получить текущего пользователя",
        "responses": {
          "500": {
            "description": "Произошла внутрення ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "Запрос выполнен успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorDto": {
        "type": "object",
        "properties": {
          "errorCode": {
            "type": "string",
            "description": "Код ошибки.",
            "nullable": true
          },
          "userMessage": {
            "type": "string",
            "description": "Сообщение для пользователя.",
            "nullable": true
          },
          "internalErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDto"
            },
            "description": "Вложенные ошибки.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Модель ошибки."
      },
...
      "ShortAdvertDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Идентификатор",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Название",
            "nullable": true
          },
          "location": {
            "type": "string",
            "description": "Геолокация объявления",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "description": "Дата создания.",
            "format": "date-time"
          },
          "isActive": {
            "type": "boolean",
            "description": "Признак актуальности объявления."
          },
          "imagesIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Изображения",
            "nullable": true
          },
          "cost": {
            "type": "number",
            "description": "Цена",
            "format": "double"
          }
        },
        "additionalProperties": false,
        "description": "Краткая информация об объявлении"
      },
...
      "ShortUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Идентификатор",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Имя пользователя",
            "nullable": true
          },
          "login": {
            "type": "string",
            "description": "Логин",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Краткая информация о пользователе"
      },
...
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Идентификатор",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Имя пользователя",
            "nullable": true
          },
          "role": {
            "type": "string",
            "description": "Роль",
            "nullable": true
          },
          "login": {
            "type": "string",
            "description": "Логин",
            "nullable": true
          },
          "adverts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShortAdvertDto"
            },
            "description": "Объявления",
            "nullable": true
          },
          "registeredTime": {
            "type": "string",
            "description": "Дата регистрации",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "Полная информация о пользователе"
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme.  \n                        Enter 'Bearer' [space] and then your token in the text input below.\n                        Example: 'Bearer secretKey'",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}